requirements.txt
# AI Email Responder Dependencies
google-generativeai==0.8.3
langchain==0.3.9
langchain-google-genai==2.0.5
langgraph==0.2.42
langchain-community==0.3.8
langchain-chroma==0.1.4
chromadb==0.5.20
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
fastmcp==0.3.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
redis==5.2.1
python-dotenv==1.0.1
pydantic==2.10.3
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.28.1
aiofiles==24.1.0
structlog==24.4.0
rich==13.9.4
typer==0.15.1 